# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Vida Refrigeração
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /service-orders:
    get:
      operationId: listServiceOrders
      tags: [service-orders]
      summary: List all service orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [open, in_progress, completed]
        - name: technician
          in: query
          required: false
          schema:
            type: string
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of service orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServiceOrder"
    post:
      operationId: createServiceOrder
      tags: [service-orders]
      summary: Create a new service order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceOrderBody"
      responses:
        "201":
          description: Service order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/{id}:
    get:
      operationId: getServiceOrder
      tags: [service-orders]
      summary: Get a service order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateServiceOrder
      tags: [service-orders]
      summary: Update a service order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceOrderBody"
      responses:
        "200":
          description: Updated service order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
    delete:
      operationId: deleteServiceOrder
      tags: [service-orders]
      summary: Delete a service order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /service-orders/{id}/start:
    post:
      operationId: startServiceOrder
      tags: [service-orders]
      summary: Inicia o cronômetro do serviço (registra hora de início)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Serviço iniciado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
  /service-orders/{id}/finish:
    post:
      operationId: finishServiceOrder
      tags: [service-orders]
      summary: Finaliza o serviço, valida obrigatórios e marca a OS como concluída
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Serviço finalizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceOrder"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        checked:
          type: boolean
      required:
        - id
        - label
        - checked
    ServiceOrder:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        clientName:
          type: string
        clientPhone:
          type: string
        clientAddress:
          type: string
        serviceType:
          type: string
          enum:
            - installation
            - maintenance
            - repair
            - cleaning
            - inspection
        technician:
          type: string
          description: "Lista legada — nome do técnico ou nomes separados por vírgula. Use 'technicians' (array)."
        technicians:
          type: array
          items:
            type: string
          description: "Lista de técnicos responsáveis pela OS (suporta múltiplos)"
        scheduledDate:
          type: string
        startTime:
          type: string
          nullable: true
          description: "Horário de início no formato HH:MM"
        endTime:
          type: string
          nullable: true
          description: "Horário de finalização prevista no formato HH:MM"
        reminderEnabled:
          type: boolean
        reminderMinutes:
          type: integer
          description: "Antecedência do lembrete em minutos (0,5,10,15,20,25,30,60)"
        status:
          type: string
          enum:
            - open
            - in_progress
            - completed
        equipmentCapacity:
          type: string
        observations:
          type: string
          nullable: true
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        photos:
          type: array
          items:
            type: string
        technicianSignature:
          type: string
          nullable: true
        clientSignature:
          type: string
          nullable: true
        serviceValue:
          type: number
          description: "Faturamento da OS em R$"
        serviceStartedAt:
          type: string
          nullable: true
          description: "Timestamp ISO de início real do serviço (cronômetro)"
        serviceCompletedAt:
          type: string
          nullable: true
          description: "Timestamp ISO de finalização real do serviço (cronômetro)"
        durationMinutes:
          type: integer
          nullable: true
          description: "Duração calculada (minutos) entre serviceStartedAt e serviceCompletedAt"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - clientName
        - clientPhone
        - clientAddress
        - serviceType
        - technician
        - technicians
        - scheduledDate
        - reminderEnabled
        - reminderMinutes
        - status
        - equipmentCapacity
        - checklist
        - photos
        - serviceValue
        - createdAt
        - updatedAt
    CreateServiceOrderBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        clientAddress:
          type: string
        serviceType:
          type: string
          enum:
            - installation
            - maintenance
            - repair
            - cleaning
            - inspection
        technician:
          type: string
          description: "Compatibilidade — preferir 'technicians'."
        technicians:
          type: array
          items:
            type: string
          description: "Lista de técnicos responsáveis (1 ou mais)"
        scheduledDate:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        reminderEnabled:
          type: boolean
        reminderMinutes:
          type: integer
        equipmentCapacity:
          type: string
        observations:
          type: string
          nullable: true
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        serviceValue:
          type: number
          description: "Faturamento da OS em R$ (>= 0)"
      required:
        - clientName
        - clientPhone
        - clientAddress
        - serviceType
        - technicians
        - scheduledDate
        - equipmentCapacity
        - checklist
    UpdateServiceOrderBody:
      type: object
      properties:
        clientName:
          type: string
        clientPhone:
          type: string
        clientAddress:
          type: string
        serviceType:
          type: string
          enum:
            - installation
            - maintenance
            - repair
            - cleaning
            - inspection
        technician:
          type: string
          description: "Compatibilidade — preferir 'technicians'."
        technicians:
          type: array
          items:
            type: string
        scheduledDate:
          type: string
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        reminderEnabled:
          type: boolean
        reminderMinutes:
          type: integer
        status:
          type: string
          enum:
            - open
            - in_progress
            - completed
        equipmentCapacity:
          type: string
        observations:
          type: string
          nullable: true
        checklist:
          type: array
          items:
            $ref: "#/components/schemas/ChecklistItem"
        photos:
          type: array
          items:
            type: string
        technicianSignature:
          type: string
          nullable: true
        clientSignature:
          type: string
          nullable: true
        serviceValue:
          type: number
...